move prop mixin
