+++ posthog/models/entity.py
@@ -1,8 +1,9 @@
 from typing import Any, Dict, List, Optional, Union
 
 from posthog.constants import TREND_FILTER_TYPE_ACTIONS, TREND_FILTER_TYPE_EVENTS
+from posthog.models.filters.mixins.common import PropertyMixin
 
-from .property import Property, PropertyMixin
+from .property import Property
 
 
 class Entity(PropertyMixin):
